Add price hook
